refactor: Move zookeeper message handling to new class
Added the zookeeper message handling to a new class so that
we can process messages the same way when we use the deadletter
queue.

Contributes to: #46
